finish migrating accept_test.go
